add zombine spawn from marine death logic
todo - make the marine not gib and have no clientsided ragdoll when a zombine spawns from his death
